refactor: Redis 인프라 패키지 이동 (stream->redis)
