+++ src/main/java/com/reservation/tablereservationservice/infrastructure/redis/ReservationPublisher.java
@@ -1,4 +1,4 @@
-package com.reservation.tablereservationservice.infrastructure.stream;
+package com.reservation.tablereservationservice.infrastructure.redis;
 
 import java.time.LocalDate;
 import java.util.List;
